backend/requirements.txt
Flask>=3.0.0
Flask-CORS>=4.0.0
Flask-SocketIO>=5.3.0
librosa>=0.10.0
python-socketio>=5.10.0
python-engineio>=4.8.0
numpy>=1.24.0
soundfile>=0.12.0
python-dotenv>=1.0.0
pytest>=7.4.0
